requirements.txt
# ================================================
# Pack 项目 - 统一依赖声明
# ================================================
# 本文件包含所有子项目的依赖汇总
# 子项目各自的 requirements.txt 仍保留（最小化修改原则）
# ================================================

# ----------------------------------------
# 核心依赖 - PyTorch & 深度学习
# ----------------------------------------
torch==2.2.2
torchvision>=0.17.0
# PyTorch CUDA 12.1 版本安装命令：
# pip install torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu121

# ----------------------------------------
# 检测模块 - YOLO & 计算机视觉
# ----------------------------------------
ultralytics==8.1.7
opencv-python>=4.13.0
pillow>=11.3.0

# ----------------------------------------
# Web 服务 - FastAPI & API
# ----------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.46.0
starlette>=1.0.0
anyio>=4.13.0

# ----------------------------------------
# 数据库 & 认证
# ----------------------------------------
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ----------------------------------------
# 数据处理
# ----------------------------------------
numpy>=1.24.0,<2.0
pandas>=2.2.2
pydantic>=2.11.7
pydantic-settings>=2.6.1
python-multipart>=0.0.27
PyYAML>=6.0.3

# ----------------------------------------
# SKU & 特征提取 - OML & 特征学习
# ----------------------------------------
# OML (Open Metric Learning) 用于特征提取和匹配
# 请参考 SKU/README.md 中的安装说明

# ----------------------------------------
# 训练 & 实验管理
# ----------------------------------------
tqdm>=4.67.1
mlflow>=2.8.0

# ----------------------------------------
# 可视化 & UI
# ----------------------------------------
gradio>=4.36.1

# ----------------------------------------
# 工具库
# ----------------------------------------
scikit-learn>=1.4.2

# ================================================
# 安装说明
# ================================================
# 1. 创建虚拟环境（推荐）
#    python -m venv .venv
#
# 2. 激活虚拟环境
#    Windows: .venv\Scripts\activate
#    Linux/Mac: source .venv/bin/activate
#
# 3. 安装 PyTorch (根据你的 CUDA 版本)
#    CUDA 12.1: pip install torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu121
#    CPU 仅: pip install torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cpu
#
# 4. 安装其他依赖
#    pip install -r requirements.txt
# ================================================
